requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
pytest>=7.4.3
python-dotenv>=1.0.0
requests>=2.31.0
pytest-cov>=4.1.0
mkdocs>=1.5.3
mkdocs-material>=9.4.6
